--- Fresco/VersionTests.swift
@@ -1,8 +1,0 @@
-@testable import FrescoFramework
-import XCTest
-
-class VersionTests: XCTestCase {
-  func test_versionCurrent() {
-    XCTAssertEqual(Version.current, "0.1.0")
-  }
-}
